wip - no more plugin, just a field on persistence
