+++ src/index.ts
@@ -5,8 +5,8 @@ import { NinePatchPlugin } from '@koreez/phaser3-ninepatch';
 import { AreaManagerPlugin } from './plugins/area-manager-plugin';
 import { ControlsPlugin } from './plugins/controls-plugin';
 import { ParallaxSpritePlugin } from './plugins/parallax-sprite-plugin';
+import { PersistencePlugin } from './plugins/persistence-plugin';
 import { PhecsPlugin } from './plugins/phecs-plugin';
-import { ProgressionPlugin } from './plugins/progression-plugin';
 import { StateRegistrarPlugin } from './plugins/state-registrar-plugin';
 
 import { ExplorationScene } from './scenes/exploration-scene';
@@ -14,7 +14,6 @@ import { DeathScene } from './scenes/death-scene';
 import { PrefabTestScene } from './scenes/prefab-test-scene';
 import { PreloadScene } from './scenes/preload-scene';
 import { HUDScene } from './scenes/hud-scene';
-import { PersistencePlugin } from './plugins/persistence-plugin';
 import { TitleScene } from './scenes/title-scene';
 import { NewGameScene } from './scenes/new-game-scene';
 import { ContinueGameScene } from './scenes/continue-game-scene';
@@ -27,7 +26,6 @@ const gameConfig = {
       { key: 'NinePatchPlugin', plugin: NinePatchPlugin, start: true },
       { key: 'ParallaxSpritePlugin', plugin: ParallaxSpritePlugin, start: true },
       { key: 'PersistencePlugin', plugin: PersistencePlugin, mapping: 'persistence', start: true },
-      { key: 'ProgressionPlugin', plugin: ProgressionPlugin, mapping: 'progression', start: true },
     ],
     scene: [
       {
