+++ src/plugins/persistence-plugin.ts
@@ -1,3 +1,6 @@
+import { Progression } from "../lib/progression";
+import { PERSISTENCE_KEYS } from "../constants/persistence-keys";
+
 type UpdateCallback<T> = (value: T) => T;
 type OnChangeCallback<T> = (value: T) => void;
 type OnChangeCleanupFn = () => void;
@@ -8,11 +11,15 @@ export class PersistencePlugin extends Phaser.Plugins.BasePlugin {
   private data: { [key: string]: any };
   private onChangeListeners: { [key: string]: OnChangeCallback<any>[] };
 
+  public progression: Progression;
+
   constructor(pluginManager: Phaser.Plugins.PluginManager) {
     super(pluginManager);
 
     this.data = {};
     this.onChangeListeners = {};
+
+    this.progression = new Progression();
   }
   
   get<T>(key: string): T {
@@ -44,6 +51,7 @@ export class PersistencePlugin extends Phaser.Plugins.BasePlugin {
   }
 
   save() {
+    this.set(PERSISTENCE_KEYS.progression, this.progression.progressionCompletion);
     localStorage.setItem(SAVE_GAME_KEY, JSON.stringify(this.data));
   }
 
@@ -51,6 +59,7 @@ export class PersistencePlugin extends Phaser.Plugins.BasePlugin {
     const savedData = localStorage.getItem(SAVE_GAME_KEY);
     if (savedData) {
       this.data = JSON.parse(savedData);
+      this.progression.setCompletionData(this.data[PERSISTENCE_KEYS.progression]);
     } else {
       throw new Error('PERSISTENCE_PLUGIN::NO_SAVED_DATA');
     }
